Id

Description

This module provides a type-wrapper around id's.

Documentation

newtype Id

Constructors

Id

Fields

_id :: Int

Resettable

class Resettable e where

Types that contain values of type Id that can be reset (set to zero).

Methods

reset :: e -> e

Reset all the Id's inside an expression.

reset :: (Generic e, GResettable (Rep e)) => e -> e

Reset all the Id's inside an expression.

class GResettable f where

Minimal complete definition

gReset

Methods

gReset :: f e -> f e

Identifiable

class Identifiable e where

Values that contain at least a value of type Id.

Methods

getId :: e -> Maybe Id

getId :: (Generic e, GIdentifiable (Rep e)) => e -> Maybe Id

class GIdentifiable f where

Minimal complete definition

gGetId

Methods

gGetId :: f e -> Maybe Id