ValExpr

Description

Interface file for Value Expressions.

Documentation

data ValExpr v

ValExpr: value expression

User can't directly construct ValExpr (such that invariants will always hold)

User can still pattern match on ValExpr using ValExprView

Overhead at run-time is zero. See https://wiki.haskell.org/Performance/Data_types#Newtypes

view :: ValExpr v -> ValExprView v

View on value expression.

eval :: Show v => ValExpr v -> Either String Const

Evaluate the provided value expression. Either the Right Constant Value is returned or a (Left) error message.

data ValExprView v

ValExprView: the public view of value expression ValExpr

Constructors

Vconst Const

Vvar v

Vequal (ValExpr v) (ValExpr v)

Vite

Fields

condition :: ValExpr v

trueBranch :: ValExpr v

falseBranch :: ValExpr v

Vnot (ValExpr v)

Vand (Set (ValExpr v))

Vdivide

Fields

dividend :: ValExpr v

divisor :: ValExpr v

Vmodulo

Fields

dividend :: ValExpr v

divisor :: ValExpr v

Vsum (FreeSum (ValExpr v))

Vproduct (FreeProduct (ValExpr v))

Vgez (ValExpr v)

Vlength (ValExpr v)

Vat

Fields

string :: ValExpr v

position :: ValExpr v

Vconcat [ValExpr v]

Vstrinre

Fields

string :: ValExpr v

regex :: ValExpr v

Vcstr CstrId [ValExpr v]

Viscstr CstrId (ValExpr v)

Vaccess CstrId Int (ValExpr v)

Vfunc FuncId [ValExpr v]

Vpredef PredefKind FuncId [ValExpr v]

data PredefKind

only needed for CNECTDEF

Constructors

AST

ASF

AXT

AXF

SSB

SSI

SSS

Constructors to create Value Expressions

Constant value

cstrConst :: Const -> ValExpr v

Create a constant as a value expression.

VarRef

cstrVar :: v -> ValExpr v

Create a variable as a value expression.

General Operators to create Value Expressions

Equal

cstrEqual :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator Equal on the provided value expressions. Preconditions are not checked.

If Then Else

cstrITE :: Eq v => ValExpr v -> ValExpr v -> ValExpr v -> ValExpr v

Apply operator ITE (IF THEN ELSE) on the provided value expressions. Preconditions are not checked.

Function Call

cstrFunc :: (Variable v, Variable w) => Map FuncId (FuncDef v) -> FuncId -> [ValExpr w] -> ValExpr w

Create a function call. Preconditions are not checked.

Boolean Operators to create Value Expressions

Not

cstrNot :: ValExpr v -> ValExpr v

Apply operator Not on the provided value expression. Preconditions are not checked.

And

cstrAnd :: Ord v => Set (ValExpr v) -> ValExpr v

Apply operator And on the provided set of value expressions. Preconditions are not checked.

Integer Operators to create Value Expressions

Sum

cstrSum :: forall v. (Ord v, Integral (ValExpr v)) => FreeSum (ValExpr v) -> ValExpr v

Apply operator sum on the provided sum of value expressions. Preconditions are not checked.

Product

cstrProduct :: forall v. (Ord v, Integral (ValExpr v)) => FreeProduct (ValExpr v) -> ValExpr v

Apply operator product on the provided product of value expressions. Be aware that division is not associative for Integer, so only use power >= 0. Preconditions are not checked.

Divide

cstrDivide :: ValExpr v -> ValExpr v -> ValExpr v

Apply operator Divide on the provided value expressions. Preconditions are not checked.

Modulo

cstrModulo :: ValExpr v -> ValExpr v -> ValExpr v

Apply operator Modulo on the provided value expressions. Preconditions are not checked.

Comparisons GEZ

cstrGEZ :: ValExpr v -> ValExpr v

Apply operator GEZ (Greater Equal Zero) on the provided value expression. Preconditions are not checked.

String Operators to create Value Expressions

Length operator

cstrLength :: ValExpr v -> ValExpr v

Apply operator Length on the provided value expression. Preconditions are not checked.

At operator

cstrAt :: ValExpr v -> ValExpr v -> ValExpr v

Apply operator At on the provided value expressions. Preconditions are not checked.

Concat operator

cstrConcat :: Eq v => [ValExpr v] -> ValExpr v

Apply operator Concat on the provided sequence of value expressions. Preconditions are not checked.

Regular Expression Operators to create Value Expressions

String in Regular Expression operator

cstrStrInRe :: ValExpr v -> ValExpr v -> ValExpr v

Apply String In Regular Expression operator on the provided value expressions. Preconditions are not checked.

Algebraic Data Type Operators to create Value Expressions

Algebraic Data Type constructor operator

cstrCstr :: CstrId -> [ValExpr v] -> ValExpr v

Apply ADT Constructor of constructor with CstrId and the provided arguments (the list of value expressions). Preconditions are not checked.

Algebraic Data Type IsConstructor function

cstrIsCstr :: CstrId -> ValExpr v -> ValExpr v

Is the provided value expression made by the ADT constructor with CstrId? Preconditions are not checked.

Algebraic Data Type Accessor

cstrAccess :: CstrId -> Int -> ValExpr v -> ValExpr v

Apply ADT Accessor of constructor with CstrId on field with given position on the provided value expression. Preconditions are not checked.

cstrPredef :: PredefKind -> FuncId -> [ValExpr v] -> ValExpr v

Create a call to a predefined function as a value expression.

Substitution of var by value

subst

Arguments

:: (Variable v, Integral (ValExpr v), Variable w, Integral (ValExpr w))

=> Map v (ValExpr v)
Map from variables to value expressions.

-> Map FuncId (FuncDef w)
Map from identifiers to their definitions, this is used to replace function calls by their bodies if all the arguments of the function are constant.

-> ValExpr v
Value expression where the substitution will take place.

-> ValExpr v

Substitute variables by value expressions in a value expression.

Preconditions are not checked.

compSubst :: (Variable v, Integral (ValExpr v), Variable w, Integral (ValExpr w)) => Map v (ValExpr w) -> Map FuncId (FuncDef v) -> ValExpr v -> ValExpr w

Substitute variables by value expressions in a value expression (change variable kind). Preconditions are not checked.

Derived Boolean operators

Or (/)

cstrOr :: Ord v => Set (ValExpr v) -> ValExpr v

Apply operator Or (\/) on the provided set of value expressions. Preconditions are not checked.

Exclusive or (|/)

cstrXor :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator Xor (\|/) on the provided set of value expressions. Preconditions are not checked.

Implies (=>)

cstrImplies :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator Implies (=>) on the provided value expressions. Preconditions are not checked.

Derived Integer operators:

Unary Plus

cstrUnaryPlus :: ValExpr v -> ValExpr v

Apply unary operator Plus on the provided value expression. Preconditions are not checked.

Unary Minus = negate single argument

cstrUnaryMinus :: Ord v => ValExpr v -> ValExpr v

Apply unary operator Minus on the provided value expression. Preconditions are not checked.

Plus = Sum of two terms

cstrPlus :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator Add on the provided value expressions. Preconditions are not checked.

Minus

cstrMinus :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator Minus on the provided value expressions. Preconditions are not checked.

Times = Product of two terms

cstrTimes :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator Times on the provided value expressions. Preconditions are not checked.

Absolute value

cstrAbs :: Ord v => ValExpr v -> ValExpr v

Apply operator Absolute value (abs) on the provided value expression. Preconditions are not checked.

Derived Integer comparisons

Less than (<)

cstrLT :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator LT (<) on the provided value expression. Preconditions are not checked.

Less Equal (<=)

cstrLE :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator LE (<=) on the provided value expression. Preconditions are not checked.

Greater Equal (>=)

cstrGE :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator GE (>=) on the provided value expression. Preconditions are not checked.

Greater Than (>)

cstrGT :: Ord v => ValExpr v -> ValExpr v -> ValExpr v

Apply operator GT (>) on the provided value expression. Preconditions are not checked.