Product

Description

Implementation for a symbolic product.

Note: Integer division is not associative, so negative occurrences should be not be used for Integers.

inspiration taken from https://hackage.haskell.org/package/multiset-0.3.3/docs/src/Data-MultiSet.html

We take 0^0 == 1

In the complexity of functions n refers to the number of distinct terms, t is the total number of terms.

Product type

type FreeProduct a = FreeMonoidX (ProductTerm a)

FreeProduct represents a symbolic product of terms of the type parameter a. The same term can occur multiple times.

newtype ProductTerm a

Terms of a free-monoids of the form:

a0 <> a1 <> ... <> an-1

where <> will be interpreted as the arithmetic multiplication of terms:

a0 * a1 * ... * an-1

Constructors

ProductTerm

Fields

factor :: a

Filter

fraction :: Ord a => FreeProduct a -> (FreeProduct a, FreeProduct a)

O(n). Partition the product into the dividend and divisor.

Product of Term and Products

multiply :: Ord a => a -> FreeProduct a -> FreeProduct a

O(log n). Multiply a product with a term.

divide :: Ord a => a -> FreeProduct a -> FreeProduct a

O(log n). Divide a product by a term.

product :: Ord a => FreeProduct a -> FreeProduct a -> FreeProduct a

O(n+m). The product of two products.

products :: Ord a => [FreeProduct a] -> FreeProduct a

The product of a list of products.

Power

power :: Ord a => Integer -> FreeProduct a -> FreeProduct a

O(n). Take the product ms to the power with the constant x.